refactor: apply trustin's feedback
